Hey so I fixed FPS for real this time Babe

+++ project/src/backend/sdl/SDLApplication.h
@@ -59,16 +59,16 @@ namespace lime {
 			bool active;
 			ApplicationEvent applicationEvent;
 			ClipboardEvent clipboardEvent;
-			Uint32 currentUpdate;
+			double currentUpdate;
 			double framePeriod;
 			DropEvent dropEvent;
 			GamepadEvent gamepadEvent;
 			JoystickEvent joystickEvent;
 			KeyEvent keyEvent;
-			Uint32 lastUpdate;
+			double lastUpdate;
 
 			MouseEvent mouseEvent;
-			Uint32 nextUpdate;
+			double nextUpdate;
 
 			RenderEvent renderEvent;
 			SensorEvent sensorEvent;
